Refactor modules. Fix #57
