+++ src/cosy/email/conf.lua
@@ -1,4 +1,4 @@
-local Default = require "cosy.configuration-layers".default
+local Default = require "cosy.configuration.layers".default
 
 Default.smtp = {
   timeout   = 2, -- seconds
