--- src/cosy/library-conf.lua
@@ -1,5 +1,0 @@
-local Default = require "cosy.configuration-layers".default
-
-Default.library = {
-  timeout = 5, -- seconds
-}
