+++ src/cosy/library/conf.lua
@@ -1,0 +1,5 @@
+local Default = require "cosy.configuration.layers".default
+
+Default.library = {
+  timeout = 5, -- seconds
+}
