+++ src/cosy/password/conf.lua
@@ -1,4 +1,4 @@
-local Default= require "cosy.configuration-layers".default
+local Default= require "cosy.configuration.layers".default
 
 Default.password = {
   time = 0.020, -- 20 milliseconds
