+++ src/cosy/string/init.lua
