+++ src/cosy/token/conf.lua
@@ -1,4 +1,4 @@
-local Default = require "cosy.configuration-layers".default
+local Default = require "cosy.configuration.layers".default
 
 Default.token = {
   algorithm = "HS512",
